requirements.txt
asgiref==3.7.2
Django==4.2.2
django-registration-redux==2.12
django-social-share==2.3.0
django-taggit==4.0.0
Pillow==9.5.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
